+++ src/ampworks/_core/__init__.py
@@ -5,6 +5,7 @@ made available at the base-level of the package.
 """
 
 from ._dataset import Dataset
+from ._headers import HeaderAliases, standardize_headers
 from ._read import read_csv, read_excel, read_table
 
 __all__ = [
@@ -12,4 +13,6 @@ __all__ = [
     'read_csv',
     'read_excel',
     'read_table',
+    'HeaderAliases',
+    'standardize_headers',
 ]